Modularize Manintenance & performance component
