+++ gui/src/components/maintenance/cortx-about.vue
@@ -80,10 +80,13 @@
 import { Component, Vue, Prop } from "vue-property-decorator";
 import { Api } from "../../services/api";
 import apiRegister from "../../services/api-register";
-import i18n from "./../../i18n";
+import i18n from "./maintenance.json";
 
 @Component({
-  name: "cortx-about"
+  name: "cortx-about",
+  i18n: {
+    messages: i18n
+  }
 })
 export default class Cortxaboutpage extends Vue {
   public data() {
