+++ gui/src/components/maintenance/maintenance.json
@@ -1,0 +1,108 @@
+{
+    "en": {
+        "systemMaintenance": {
+            "page-title": "System maintenance is working",
+            "start-service": "Start service:",
+            "stop-service": "Stop service:",
+            "shutdown": "Shutdown:",
+            "apply-btn": "Apply",
+            "system-info-dialog-message": "The services have been restarted. You will have to login again.",
+            "fetching-system-message": "Fetching system resources...",
+            "error-message-onclose": "No response, please refresh the page.",
+            "confirm-message-node-stop": "Are you sure, you want to stop the node?",
+            "danger-severity": "danger",
+            "info-severity": "info",
+            "start-action-method": "start",
+            "stop-action-method": "stop",
+            "shutdown-action-method": "shutdown",
+            "confirm-message-node-start": "Are you sure, you want to start the node?",
+            "confirm-message-node-shutdown": "Are you sure, you want to shutdown the node?",
+            "confirm-message-shutdown-sub": "Please note that if you shutdown the node then the entire application will stop",
+            "shutdown-note": "Note: {node} is in shutdown state. You need to start it manually."
+        },
+        "maintenance": {
+            "auditLog": "Audit log",
+            "downloadOrView": "Download or view audit logs for the selected time period.",
+            "component": "Component",
+            "timePeriod": "Time period",
+            "download": "Download",
+            "view": "View",
+            "logs": "Logs",
+            "nodeReplacement": "Node replacement",
+            "lastNodeReplacementStatus": "Last node replacement status",
+            "refresh": "Refresh",
+            "replacementOf": "Replacement of",
+            "IPHostnameOptional": "IP/Hostname (optional)",
+            "invalidIPHostname": "Invalid IP/Hostname.",
+            "portOptional": "Port (optional)",
+            "invalidSSH": "SSH port is not valid.",
+            "replaceNode": "Replace node",
+            "allNodesOnline": "All nodes are online.",
+            "allNodesOnlineNote": "Note: Node replacement is only available when any one of the nodes is in shutdown state. In order to complete node replacement please shutdown the node from system maintenance page.",
+            "updateSoftware": "Update software",
+            "updateSoftwareFile": "Update the software. Upload the software bundle file (.iso) received from Seagate.",
+            "updateSoftwareClick": "Click Start update once the bundle file is uploaded successfully. After the software update, you must login again.",
+            "lastUpdateStatus": "Last update status",
+            "lastUpdateVersion": "Last update version",
+            "lastUpdateDescription": "Last update description",
+            "uploadNewSoftwareFile": "Upload new software file",
+            "startUpdate": "Start update",
+            "invalidFile": "Invalid file.",
+            "upload": "Upload",
+            "cancel": "Cancel"
+        }
+    },
+    "ru": {
+        "systemMaintenance": {
+            "page-title": "System maintenance",
+            "start-service": "Start service:",
+            "stop-service": "Stop service:",
+            "shutdown": "Shutdown:",
+            "apply-btn": "Apply",
+            "system-info-dialog-message": "The services have been restarted. You will have to login again.",
+            "fetching-system-message": "Fetching system resources...",
+            "error-message-onclose": "No response, please refresh the page.",
+            "confirm-message-node-stop": "Are you sure, you want to stop the node?",
+            "danger-severity": "danger",
+            "info-severity": "info",
+            "start-action-method": "start",
+            "stop-action-method": "stop",
+            "shutdown-action-method": "shutdown",
+            "confirm-message-node-start": "Are you sure, you want to start the node?",
+            "confirm-message-node-shutdown": "Are you sure, you want to shutdown the node?",
+            "confirm-message-shutdown-sub": "Please note that if you shutdown the node then the entire application will stop",
+            "shutdown-note": "Note: {node} is in shutdown state. You need to start it manually."
+        },
+        "maintenance": {
+            "auditLog": "Audit log",
+            "downloadOrView": "Download or view audit logs for the selected time period.",
+            "component": "Component",
+            "timePeriod": "Time period",
+            "download": "Download",
+            "view": "View",
+            "logs": "Logs",
+            "nodeReplacement": "Node replacement",
+            "lastNodeReplacementStatus": "Last node replacement status",
+            "refresh": "Refresh",
+            "replacementOf": "Replacement of",
+            "IPHostnameOptional": "IP/Hostname (optional)",
+            "invalidIPHostname": "Invalid IP/Hostname.",
+            "portOptional": "Port (optional)",
+            "invalidSSH": "SSH port is not valid.",
+            "replaceNode": "Replace node",
+            "allNodesOnline": "All nodes are online.",
+            "allNodesOnlineNote": "Note: Node replacement is only available when any one of the nodes is in shutdown state. In order to complete node replacement please shutdown the node from system maintenance page.",
+            "updateSoftware": "Update software",
+            "updateSoftwareFile": "Update the software. Upload the software bundle file (.iso) received from Seagate.",
+            "updateSoftwareClick": "Click Start update once the bundle file is uploaded successfully. After the software update, you must login again.",
+            "lastUpdateStatus": "Last update status",
+            "lastUpdateVersion": "Last update version",
+            "lastUpdateDescription": "Last update description",
+            "uploadNewSoftwareFile": "Upload new software file",
+            "startUpdate": "Start update",
+            "invalidFile": "Invalid file.",
+            "upload": "Upload",
+            "cancel": "Cancel"
+        }
+    }
+}
